+++ internal/markdownmode/spec.go
@@ -2,7 +2,6 @@ package markdownmode
 
 import (
 	"github.com/dogmatiq/ferrite/variable"
-	"github.com/mattn/go-runewidth"
 )
 
 func (r *renderer) renderSpecs() {
@@ -94,47 +93,4 @@ func (r schemaRenderer) VisitOther(variable.Other) {
 func (r *renderer) renderUndefinedFailureWarning() {
 	r.line("If left undefined the application will print usage information to `STDERR` then")
 	r.line("exit with a non-zero exit code.")
-}
-
-func (r *renderer) renderExamples(s variable.Spec) {
-	r.line("```bash")
-
-	width := 0
-	for _, eg := range s.Examples() {
-		w := runewidth.StringWidth(eg.Canonical.Quote())
-		if w > width {
-			width = w
-		}
-	}
-
-	for _, eg := range s.Examples() {
-		comment := ""
-		if variable.IsDefault(s, eg.Canonical) {
-			comment = "(default)"
-		}
-		if eg.Description != "" {
-			if comment != "" {
-				comment += " "
-			}
-			comment += eg.Description
-		}
-
-		if len(comment) == 0 {
-			r.line(
-				"export %s=%s",
-				s.Name(),
-				eg.Canonical.Quote(),
-			)
-		} else {
-			r.line(
-				"export %s=%-*s # %s",
-				s.Name(),
-				width,
-				eg.Canonical.Quote(),
-				comment,
-			)
-		}
-	}
-
-	r.line("```")
 }