+++ internal/markdownmode/spec_test.go
@@ -35,7 +35,7 @@ var _ = Describe("func Run()", func() {
 			expect, err := os.ReadFile(filepath.Join("testdata", "markdown", file))
 			Expect(err).ShouldNot(HaveOccurred())
 
-			actual := Run("<app>", reg, false)
+			actual := Run("<app>", reg, WithoutUsageExamples())
 			ExpectWithOffset(1, actual).To(EqualX(string(expect)))
 		},
 		Entry(