+++ markdownexample_test.go
@@ -26,8 +26,8 @@ func ExampleInit_markdownUsage() {
 	//
 	// **There do not appear to be any environment variables.**
 	//
-	// The application may consume other undocumented environment variables; this
-	// document only shows those variables defined using [Ferrite].
+	// ⚠️ The application may consume other undocumented environment variables; this
+	// document only shows those variables declared using [Ferrite].
 	//
 	// <!-- references -->
 	//